refactor(workspace): move state observers out of instrumentation
Per CONVENTIONS.md, src/instrumentation is telemetry-only. The pure
detection code has no telemetry or logger imports, so move TransitionTracker,
WorkspaceStateObserver, WorkspaceAgentObserver, and the transition types to
src/workspace/observers.ts. The record* functions stay in instrumentation and
import the transition types from there. Observer/tracker tests move to
test/unit/workspace/observers.test.ts.
